requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-filter
python-decouple
psycopg2
psycopg2-binary
Pillow